glass: Relocate forms to dedicated directory
Relocate existing forms to a dedicated directory because they are used at different places within the UI.

Additional actions:
- Modified the look of the dialog.
- Changed color theme from yellow/white to yellow/black according to EOS.
- Add missing TEXT() to be able to extract the messages automatically for i18n.
