+++ src/glass/src/app/core/core.module.ts
@@ -9,6 +9,7 @@ import { DashboardModule } from '~/app/core/dashboard/dashboard.module';
 import { BlankLayoutComponent } from '~/app/core/layouts/blank-layout/blank-layout.component';
 import { InstallerLayoutComponent } from '~/app/core/layouts/installer-layout/installer-layout.component';
 import { MainLayoutComponent } from '~/app/core/layouts/main-layout/main-layout.component';
+import { CephfsModalComponent } from '~/app/core/modals/cephfs/cephfs-modal.component';
 import { NavigationBarComponent } from '~/app/core/navigation-bar/navigation-bar.component';
 import { NavigationBarItemComponent } from '~/app/core/navigation-bar/navigation-bar-item/navigation-bar-item.component';
 import { TopBarComponent } from '~/app/core/top-bar/top-bar.component';
@@ -22,7 +23,8 @@ import { SharedModule } from '~/app/shared/shared.module';
     NavigationBarComponent,
     InstallerLayoutComponent,
     BlankLayoutComponent,
-    NavigationBarItemComponent
+    NavigationBarItemComponent,
+    CephfsModalComponent
   ],
   imports: [
     BlockUIModule.forRoot(),
