+++ src/glass/src/app/core/modals/cephfs/cephfs-modal.component.scss
@@ -1,0 +1,38 @@
+@import 'styles.scss';
+
+.glass-cephfs-modal {
+  .warning {
+    @extend .glass-color-theme-warn;
+    margin-bottom: 8px;
+
+    .mat-icon {
+      margin-right: 16px;
+    }
+  }
+
+  ::ng-deep .mat-form-field {
+    &.glass-form-field-no-underline {
+      .mat-form-field-underline {
+        visibility: hidden;
+      }
+    }
+
+    &.glass-form-field-hint {
+      margin-bottom: 10px;
+    }
+
+    &.mat-form-field-invalid {
+      .mat-input-element {
+        color: $eos-bc-red-500;
+      }
+      .mat-slider {
+        .mat-slider-track-fill,
+        .mat-slider-thumb-label,
+        .mat-slider-thumb,
+        .mat-slider-focus-ring {
+          background-color: $eos-bc-red-500;
+        }
+      }
+    }
+  }
+}
