+++ src/glass/src/app/core/modals/cephfs/cephfs-modal.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, ValidationErrors, ValidatorFn, Validators } from '@angular/forms';
 import { MatDialogRef } from '@angular/material/dialog';
 import { MatSliderChange } from '@angular/material/slider';
+import { marker as TEXT } from '@biesbjerg/ngx-translate-extract-marker';
 import * as _ from 'lodash';
 import { BlockUI, NgBlockUI } from 'ng-block-ui';
 import { finalize } from 'rxjs/operators';
@@ -27,7 +28,7 @@ export class CephfsModalComponent implements OnInit {
   public showWarning = false;
   public showWarningText: string[] = [];
 
-  private constraints: Constraints|undefined = undefined;
+  private constraints: Constraints | undefined = undefined;
 
   public constructor(
     private dialogRef: MatDialogRef<CephfsModalComponent>,
@@ -78,14 +79,16 @@ export class CephfsModalComponent implements OnInit {
       return;
     }
     const values = this.formGroup.value;
-    this.blockUI.start('Please wait, assessing service specifications ...');
+    this.blockUI.start(TEXT('Please wait, assessing service specifications ...'));
     this.services
       .checkRequirements(values.requiredSpace, values.replicas)
       .pipe(finalize(() => this.blockUI.stop()))
       .subscribe({
         next: (result: CheckRequirementsReply) => {
           if (!result.feasible) {
-            this.notification.show('Service creation requirements not met', { type: 'error' });
+            this.notification.show(TEXT('Service creation requirements not met.'), {
+              type: 'error'
+            });
             return;
           }
           this.createService();
@@ -107,14 +110,14 @@ export class CephfsModalComponent implements OnInit {
       if (values.replicas > maxReplicas) {
         this.showWarning = true;
         this.showWarningText.push(
-          `Current deployment can only guarantee ${maxReplicas} replicas`
+          TEXT(`Current deployment can only guarantee ${maxReplicas} replicas.`)
         );
       }
 
       if (numHosts === 1) {
         this.showWarning = true;
         this.showWarningText.push(
-          'Current deployment has a single host and can\'t tolerate failures'
+          TEXT('Current deployment has a single host and can\'t tolerate failures.')
         );
       }
     }
@@ -122,7 +125,7 @@ export class CephfsModalComponent implements OnInit {
 
   private createService(): void {
     const values = this.formGroup.value;
-    this.blockUI.start('Please wait, deploying CephFS service ...');
+    this.blockUI.start(TEXT('Please wait, deploying CephFS service ...'));
     this.services
       .create(values.name, 'cephfs', values.requiredSpace, values.replicas)
       .pipe(finalize(() => this.blockUI.stop()))
@@ -130,9 +133,9 @@ export class CephfsModalComponent implements OnInit {
         next: (result: CreateServiceReply) => {
           let success = false;
           if (!result.success) {
-            this.notification.show('Failed to create service', { type: 'error' });
+            this.notification.show(TEXT('Failed to create service.'), { type: 'error' });
           } else {
-            this.notification.show('Service successfully created');
+            this.notification.show(TEXT('Service successfully created.'));
             success = true;
           }
           this.dialogRef.close(success);
