--- src/glass/src/app/pages/deployment-page/cephfs-modal/cephfs-modal.component.html
@@ -1,127 +1,0 @@
-<div fxLayout="row"
-     fxLayoutAlign="start start"
-     fxLayoutGap="16px">
-  <h1 mat-dialog-title>CephFS</h1>
-  <div *ngIf="showWarning"
-       fxLayout="column"
-       fxLayoutAlign="start start"
-       style="width: 100%;">
-    <span *ngFor="let warning of showWarningText"
-          fxLayout="row"
-          fxLayoutAlign="start center"
-          fxLayoutGap="5px"
-          style="font-size: 12px; width: 100%; color: black"
-          class="glass-color-theme-warn">
-      <mat-icon svgIcon="mdi:alert" style="transform: scale(0.6)"></mat-icon>
-      <span>{{warning}}</span>
-    </span>
-  </div>
-</div>
-<mat-dialog-content fxLayout="column"
-                    fxLayoutAlign="none">
-  <form [formGroup]="formGroup"
-        class="mat-typography"
-        novalidate
-        fxLayout="column"
-        fxLayoutAlign="none">
-    <div fxLayout="row"
-         fxLayoutAlign="space-between none"
-         fxLayoutGap="16px">
-      <mat-form-field fxFlex>
-        <mat-label>Available Capacity</mat-label>
-        <input matInput
-               type="text"
-               [value]="formGroup.value.availableSpace | bytesToSize"
-               readonly>
-      </mat-form-field>
-      <mat-form-field fxFlex>
-        <mat-label>Reserved Capacity</mat-label>
-        <input matInput
-               type="text"
-               [value]="formGroup.value.reservedSpace | bytesToSize"
-               readonly>
-      </mat-form-field>
-      <mat-form-field [ngClass]="{'mat-form-field-invalid': formGroup.hasError('overBudget', 'requiredSpace')}"
-                      fxFlex>
-        <mat-label>Raw Required Capacity</mat-label>
-        <input matInput
-               type="text"
-               [value]="formGroup.value.rawRequiredSpace | bytesToSize"
-               readonly>
-      </mat-form-field>
-    </div>
-    <mat-form-field fxFlex>
-      <mat-label>Service Name</mat-label>
-      <input matInput
-             type="text"
-             formControlName="name"
-             required>
-      <mat-error *ngIf="formGroup.invalid">
-        <span *ngIf="formGroup.hasError('required', 'name')">
-          This field is required.
-        </span>
-      </mat-error>
-    </mat-form-field>
-    <mat-form-field class="glass-form-field-no-underline glass-form-field-hint"
-                    fxFlex>
-      <mat-label>Number Of Replicas</mat-label>
-      <input #replicasInput
-             matInput
-             type="number"
-             formControlName="replicas"
-             hidden>
-      <mat-slider color="primary"
-                  [value]="replicasInput.value"
-                  (input)="onReplicasChange($event)"
-                  min="1"
-                  max="3"
-                  step="1"
-                  thumbLabel="true"
-                  fxFlex>
-      </mat-slider>
-      <mat-hint>
-        <span>{{formGroup.get('replicas')?.value | redundancyLevel:'long'}}</span>
-      </mat-hint>
-    </mat-form-field>
-    <div fxLayout="row"
-         fxLayoutAlign="space-between center"
-         fxLayoutGap="16px">
-      <mat-form-field class="glass-form-field-no-underline"
-                      [ngClass]="{'mat-form-field-invalid': formGroup.hasError('overBudget', 'requiredSpace')}"
-                      fxFlex="80">
-        <mat-label>Estimated Required Capacity</mat-label>
-        <input #requiredSpaceInput
-               matInput
-               type="number"
-               formControlName="requiredSpace"
-               hidden>
-        <mat-slider color="primary"
-                    [value]="requiredSpaceInput.value"
-                    (input)="onRequiredSpaceChange($event)"
-                    min="0"
-                    [max]="formGroup.value.availableSpace"
-                    step="1"
-                    fxFlex>
-        </mat-slider>
-      </mat-form-field>
-      <mat-form-field [ngClass]="{'mat-form-field-invalid': formGroup.hasError('overBudget', 'requiredSpace')}"
-                      fxFlex="20">
-        <input matInput
-               type="text"
-               [value]="formGroup.value.requiredSpace | bytesToSize"
-               readonly>
-      </mat-form-field>
-    </div>
-  </form>
-</mat-dialog-content>
-<mat-dialog-actions>
-  <button mat-button
-          [mat-dialog-close]="false">
-    <span translate>Cancel</span>
-  </button>
-  <glass-submit-button [form]="formGroup"
-                       (buttonClick)="onSubmit()"
-                       cdkFocusInitial>
-    <span translate>Create</span>
-  </glass-submit-button>
-</mat-dialog-actions>
