--- src/glass/src/app/pages/deployment-page/cephfs-modal/cephfs-modal.component.scss
@@ -1,25 +1,0 @@
-::ng-deep .mat-form-field {
-  &.glass-form-field-no-underline {
-    .mat-form-field-underline {
-      visibility: hidden;
-    }
-  }
-
-  &.glass-form-field-hint {
-    margin-bottom: 10px;
-  }
-
-  &.mat-form-field-invalid {
-    .mat-input-element {
-      color: #dc3545;
-    }
-    .mat-slider {
-      .mat-slider-track-fill,
-      .mat-slider-thumb-label,
-      .mat-slider-thumb,
-      .mat-slider-focus-ring {
-        background-color: #dc3545;
-      }
-    }
-  }
-}
