+++ src/glass/src/app/pages/deployment-page/deployment-page.component.html
@@ -128,11 +128,11 @@
          fxFlexFill>
       <div>
         <mat-icon *ngIf="deploymentSuccessful"
-                  class="glass-icon-large"
+                  class="glass-icon-4x"
                   svgIcon="mdi:check-circle-outline">
         </mat-icon>
         <mat-icon *ngIf="!deploymentSuccessful"
-                  class="glass-icon-large"
+                  class="glass-icon-4x"
                   svgIcon="mdi:close-circle-outline">
         </mat-icon>
       </div>
