+++ src/glass/src/app/pages/not-found-page/not-found-page.component.html
@@ -2,7 +2,7 @@
      fxLayoutAlign="center center">
   <div fxLayout="column"
        fxLayoutAlign="center center">
-    <mat-icon class="glass-icon-large"
+    <mat-icon class="glass-icon-4x"
               svgIcon="mdi:fishbowl-outline">
     </mat-icon>
     <h1 translate>Sorry, we could not find what you are looking for.</h1>
