+++ src/glass/src/app/pages/pages.module.ts
@@ -8,7 +8,6 @@ import { CoreModule } from '~/app/core/core.module';
 import { MaterialModule } from '~/app/material.modules';
 import { BootstrapPageComponent } from '~/app/pages/bootstrap-page/bootstrap-page.component';
 import { DashboardPageComponent } from '~/app/pages/dashboard-page/dashboard-page.component';
-import { CephfsModalComponent } from '~/app/pages/deployment-page/cephfs-modal/cephfs-modal.component';
 import { DeploymentPageComponent } from '~/app/pages/deployment-page/deployment-page.component';
 import { InstallModePageComponent } from '~/app/pages/install-mode-page/install-mode-page.component';
 import { NotFoundPageComponent } from '~/app/pages/not-found-page/not-found-page.component';
@@ -25,7 +24,6 @@ import { SharedModule } from '~/app/shared/shared.module';
     InstallModePageComponent,
     WelcomePageComponent,
     NotFoundPageComponent,
-    CephfsModalComponent,
     RegisterPageComponent,
     ServicesPageComponent
   ],
