+++ src/glass/src/defaults.scss
@@ -1,6 +1,7 @@
 $glass-color-transparent: transparent;
 $glass-color-transparent-black: rgba(0, 0, 0, 0.54);
 $glass-color-white: white;
+$glass-color-black: black;
 
 $eos-bc-red-500: #dc3545;
 $eos-bc-pine-500: #0c322c;
